refactor, so we have an `Arrow` function
